fix(documentation-v7): fix alert stories
