+++ packages/documentation-v7/src/stories/components/alert/alert.docs.mdx
@@ -7,7 +7,10 @@ import { BADGE } from '../../../../.storybook/constants';
 
 # Alert
 
-Provide contextual feedback messages for typical user actions with the handful of available and flexible alert messages.<br/>
+<div class="lead">
+Provide contextual feedback messages for typical user actions with the handful of available and flexible alert messages.
+</div>
+
 Alerts are available for any length of text, as well as an optional close button.
 By default, alerts are rendered with a contextually appropriate icon.
